refactor(yaml): extract flow style and block scalar parsing into subsystem extension files
